refactor(objects): Add remaining files to major object refactor
