+++ pyartifactory/utils.py
@@ -1,6 +1,16 @@
+# Copyright (c) 2019 Ananias
+# Copyright (c) 2023 Helio Chissini de Castro
+#
+# Licensed under the MIT license: https://opensource.org/licenses/MIT
+# Permission is granted to use, copy, modify, and redistribute the work.
+# Full license information available in the project LICENSE file.
+#
+# SPDX-License-Identifier: MIT
+
 """
 Definition of all utils.
 """
+from __future__ import annotations
 
 from typing import Any
